refactor: move `pdf.py` and `image.py` from the “pdf_image” directory to just the “partition” directory to make them coherent with the importation code of other types of documents
